experiments/benchmark/tasks/feature_addition/repos/feat05_django_jwt/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
pytest>=7.4.0
pytest-django>=4.5.0
